# swagger.yaml
openapi: 3.0.4
info:
  title: LAPEN Agenda API
  description: Tennis Court Management System with Betting Integration
  version: 1.0.0
  contact:
    email: lapen.ptc@gmail.com
servers:
  - url: http://localhost:5001
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminSession:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Error message description"
...
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
...
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request parameters"
paths:
...
  /api/statistics/match-result:
    post:
      tags:
        - Statistics
      summary: Add match result
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - winner_name
              properties:
                schedule_id:
                  type: integer
                ranking_match_id:
                  type: integer
                winner_name:
                  type: string
                player1_sets:
                  type: integer
                player2_sets:
                  type: integer
                player1_games:
                  type: integer
                player2_games:
                  type: integer
      responses:
        '201':
          description: Result added successfully
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/statistics/player:
    get:
      tags:
        - Statistics
      summary: Get player statistics
      parameters:
        - name: player1
          in: query
          required: true
          schema:
            type: string
        - name: player2
          in: query
          schema:
            type: string
        - name: match_type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Player statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_matches:
                    type: integer
                  wins:
                    type: integer
                  losses:
                    type: integer
                  sets_won:
                    type: integer
                  sets_lost:
                    type: integer
                  games_won:
                    type: integer
                  games_lost:
                    type: integer
                  head_to_head:
                    type: object
                  matches:
                    type: array
  /api/statistics/past-matches:
    get:
      tags:
        - Statistics
      summary: Get past matches without results
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of past matches
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/statistics/opponents/{player_name}:
    get:
      tags:
        - Statistics
      summary: Get opponents for a specific player
      parameters:
        - name: player_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of opponents
          content:
            application/json:
              schema:
                type: object
                properties:
                  opponents:
                    type: array
                    items:
                      type: string
...